backend/functions/requirements.txt
firebase-functions==0.1.2
firebase-admin==6.1.0
python-dotenv==1.0.0